fix loadBanner to work correctly with render prop
With the addition of the render prop, it became possible for loadBanner() to be called before we have a ref to banner view. With this fix, if the loadBanner() is called before ref is set, we wait for ref to be set before using it.

This commit also changes RNPublisherBanner to only call loadBanner() on mount if render prop isn't provided. If render is provided, the parent component which provided `render` prop will be passed `loadBanner` as argument, and it is its responsiblity to call it at the appropriate time.

# Conflicts:
#	RNPublisherBanner.js